fixup! Database: Replace enum handling in alembic

+++ lib/rucio/db/sqla/migrate_repo/versions/45378a1e76a8_create_collection_replica_table.py
@@ -75,5 +75,6 @@ def downgrade():
     Downgrade the database to the previous revision
     '''
 
-    try_drop_constraint('COLLECTION_REPLICAS_STATE_CHK', 'collection_replicas', type_='check')
-    drop_table('collection_replicas')
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        try_drop_constraint('COLLECTION_REPLICAS_STATE_CHK', 'collection_replicas', type_='check')
+        drop_table('collection_replicas')
